+++ sertit/rasters.py
@@ -1006,6 +1006,22 @@ def _3d_to_2d(function):
 
 
 @__read__any_raster_to_rio_ds
+def get_window(ds: AnyRasterType, window: Any):
+    """
+    Get a window from any type of input
+
+    Args:
+        ds (AnyRasterType): Path to the raster, its dataset, its :code:`xarray` or a tuple containing its array and metadata
+        window (Any): Anything that can be returned as a window. In case of iterable, assumption is made it's geographic bounds. For pixel, please provide a Window directly.
+
+    Returns:
+        Window: Rasterio window
+
+    """
+    return rasters_rio.get_window(ds, window)
+
+
+@__read__any_raster_to_rio_ds
 def read(
     ds: AnyRasterType,
     resolution: tuple | list | float = None,
@@ -1069,7 +1085,7 @@ def read(
 
     # Manage window
     if window is not None:
-        window = rasters_rio.get_window(ds, window)
+        window = get_window(ds, window)
 
     # Get new height and width
     new_height, new_width, do_resampling = rasters_rio.get_new_shape(