+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - **BREAKING CHANGE: Rename `CustomEncoder` into `CustomJsonEncoder` (shouldn't affect anyone)**
 - **ENH: Add a YAML save and read on JSON's model (with custom dumper)**
+- **ENH: Propagate `rasters_rio.get_window` into `rasters`**
 
 ## 1.52.0 (2026-04-15)